Fix evaluation of sequences
Before calling evaluateSequence recursively, we need to remove the
statement that we just evaluated from the list of statements.

Before returning from evaluateSequence, we need to add back the
statement that we just evaluated, into the list of statements.
This allows this statement to be re-evaluated in the "next world".